pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.15",
    "langchain-openai>=1.1.14",
    "langchain-tavily>=0.2.18",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
    "ruff>=0.15.11",
    "tavily-python>=0.7.23",
]
